requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.43.0
sqlalchemy==2.0.49
jinja2==3.1.6
python-multipart==0.0.22
aiofiles==25.1.0
bleach==6.2.0
pytest==9.0.2
pytest-cov==7.1.0
httpx==0.28.1
pytest-playwright==0.7.2
playwright==1.58.0
boto3==1.38.34

# Character import pipeline (Phase 3+)
python-magic==0.4.27
striprtf==0.0.29
pypdf==6.10.2
pypdfium2==5.7.0
python-docx==1.2.0
openpyxl==3.1.5
xlrd==1.2.0
odfpy==1.4.1
beautifulsoup4==4.14.3

# Character art pipeline
Pillow==12.2.0
opencv-python-headless==4.13.0.92
